% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/epididact.R
\name{HIR}
\alias{HIR}
\title{Model HIR: Healthy - Infected - Removed}
\usage{
HIR(times, y, parms)
}
\arguments{
\item{x}{a numeric}
}
\value{
something
}
\description{
#' COUPLED DIFFERENTIAL EQUATION MODELS TO STUDY DISEASE DEVELOPMENT
BUILDING ON EXAMPLES PRESENTED IN MADDEN ET AL. 2007, CHAPTER 5
}

