% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/epididact.R
\name{HLIR}
\alias{HLIR}
\title{Model HLIR: Healthy - Latent - Infected - Removed}
\usage{
HLIR(times, y, parms)
}
\arguments{
\item{times}{}

\item{y}{}

\item{parms}{}
}
\value{
something
}
\description{
Coupled differential equation models to study disease development
building on examples presented in madden et al. 2007, chapter 5
}

