% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/epididact.R
\name{plot2gomp}
\alias{plot2gomp}
\title{plot2gomp}
\usage{
plot2gomp(y0_1, y0_2, r_1, r_2, maxt)
}
\arguments{
\item{y0_1}{Epidemic 1: initial inoculum}

\item{y0_2}{Epidemic 2: initial inoculum}

\item{r_1}{Epidemic 1: rate of progress}

\item{r_2}{Epidemic 2: rate of progress}

\item{maxt}{epidemic duration (days)}
}
\value{
two gompertz curves
}
\description{
Comparison of two epidemics (gompertz models) in the same plot
}
\examples{
plot2gomp(0.01, 0.01, 0.05, 0.025, 100)
}

