% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/epididact.R
\name{plot2gomp3J}
\alias{plot2gomp3J}
\title{plot2gomp3J}
\usage{
plot2gomp3J(a_1, a_2, b_1, b_2, r_1, r_2, maxt)
}
\arguments{
\item{a_1}{Epidemic 1: curve assintote}

\item{a_2}{Epidemic 2: curve assintote}

\item{r_1}{Epidemic 1: rate of progress}

\item{r_2}{Epidemic 2: rate of progress}

\item{maxt}{epidemic duration (days)}

\item{y0_1}{Epidemic 1: time of start of epidemic}

\item{y0_2}{Epidemic 2: time of start of epidemic}
}
\value{
Three gompertz curves
}
\description{
Comparison of three epidemics (gompertz model) at the same plot.
Alternative parametrization
}
\examples{
plot2gomp3J(1, 1, 3, 6, 0.4, 0.4, 50)
}

