% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/epididact.R
\name{plot2gomp3M}
\alias{plot2gomp3M}
\title{plot2gomp3M}
\usage{
plot2gomp3M(a_1, a_2, y0_1, y0_2, r_1, r_2, maxt)
}
\arguments{
\item{a_1}{Epidemic 1: curve assintote}

\item{a_2}{Epidemic 2: curve assintote}

\item{y0_1}{Epidemic 1: initial inoculum}

\item{y0_2}{Epidemic 2: initial inoculum}

\item{r_1}{Epidemic 1: rate of progress}

\item{r_2}{Epidemic 2: rate of progress}

\item{maxt}{epidemic duration (days)}
}
\value{
Three gompertz curves
}
\description{
Comparison of three epidemics (gompertz model) at the same plot. (from Madden, 2007)
}
\examples{
plot2gomp3M(1, 0.7, 0.001, 0.001, 0.4, 0.4, 50)
}

