% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/epididact.R
\name{plot2linear}
\alias{plot2linear}
\title{2 Parameters ####
plot2linear
Comparison of two epidemics (linear models) in the same plot}
\usage{
plot2linear(y0_1, y0_2, r_1, r_2, maxt)
}
\arguments{
\item{y0_1}{Epidemic 1: initial inoculum}

\item{y0_2}{Epidemic 2: initial inoculum}

\item{r_1}{Epidemic 1: rate of progress}

\item{r_2}{Epidemic 2: rate of progress}

\item{maxt}{epidemic duration (days)}
}
\value{
two rect lines
}
\description{
2 Parameters ####
plot2linear
Comparison of two epidemics (linear models) in the same plot
}
\examples{
plot2linear(0.01, 0.01, 0.03, 0.02, 30)
}

