% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/epididact.R
\name{plot2logis3M}
\alias{plot2logis3M}
\title{plot2logis3M}
\usage{
plot2logis3M(a_1, a_2, y0_1, y0_2, r_1, r_2, maxt)
}
\arguments{
\item{y0_1}{Epidemic 1: initial inoculum}

\item{y0_2}{Epidemic 2: initial inoculum}

\item{r_1}{Epidemic 1: rate of progress}

\item{r_2}{Epidemic 2: rate of progress}

\item{maxt}{epidemic duration (days)}
}
\value{
Two logistic curves
}
\description{
Comparison of three epidemics (logisic model) at the same plot.
from Madden, 2007: a/(1+((a/y0)-1)*exp(-r*x)),
}
\examples{
plot2logis3M(1, 0.7, 0.001, 0.001, 0.4, 0.4, 50)
}

