% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/epididact.R
\name{plot2mono}
\alias{plot2mono}
\title{plot2mono}
\usage{
plot2mono(y0_1, y0_2, r_1, r_2, maxt)
}
\arguments{
\item{y0_1}{intensidad de inóculo inicial de la epidemia 1}

\item{y0_2}{intensidad de inóculo inicial de la epidemia 2}

\item{r_1}{tasa de progreso de epidemia 1}

\item{r_2}{tasa de progreso de epidemia 2}

\item{maxt}{tiempo de epidemia}
}
\value{
Dos curvas monomolecular
}
\description{
Compara dos epidemias monomolecular en el mismo grafico
}

