% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/epididact.R
\name{plot2mono3M}
\alias{plot2mono3M}
\title{plot2mono3M}
\usage{
plot2mono3M(a_1, a_2, y0_1, y0_2, r_1, r_2, maxt)
}
\arguments{
\item{a_1}{Epidemic 1: assintote}

\item{a_2}{Epidemic 2: assintote}

\item{y0_1}{Epidemic 1: initial inoculum}

\item{y0_2}{Epidemic 2: initial inoculum}

\item{r_1}{Epidemic 1: rate of progress}

\item{r_2}{Epidemic 2: rate of progress}

\item{maxt}{epidemic duration (days)}
}
\value{
Three gompertz curves
}
\description{
Comparison of three epidemics (monomolecular model) at the same plot. (from Madden, 2007)
Adaptado de Madden, 2007 -> y = 1 - (1-y0)*exp(-r * x)
}
\examples{
plot2mono3M(
0.8, 0.8,  # a:  Assintote
0.4, 0.1,  # y0: Initial inoculum (Intercept)
0.25, 0.25,# r:  progress rate
20)        # maxt: epidemic duration
}

