% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/epididact.R
\name{plot3logis}
\alias{plot3logis}
\title{plot2logis}
\usage{
plot3logis(y0_1, y0_2, y0_3, r_1, r_2, r_3, maxt)
}
\arguments{
\item{y0_1}{intensidad de inóculo inicial de la epidemia 1}

\item{y0_2}{intensidad de inóculo inicial de la epidemia 2}

\item{y0_3}{intensidad de inóculo inicial de la epidemia 3}

\item{r_1}{tasa de progreso de epidemia 1}

\item{r_2}{tasa de progreso de epidemia 2}

\item{maxt}{tiempo de epidemia}

\item{r_1}{tasa de progreso de epidemia 3}
}
\value{
Tres curvas logisticas
}
\description{
Compara dos epidemias logisticas en el mismo grafico
}

