% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/epididact.R
\name{plot3model}
\alias{plot3model}
\title{plot3model}
\usage{
plot3model(y0, r, maxt)
}
\arguments{
\item{y0}{intensidad de inóculo inicial}

\item{r}{tasa de progreso}

\item{maxt}{tiempo de epidemia}
}
\value{
tres curvas L, M y G en distintos gráficos
}
\description{
Compara modelos monomolecular, logistico y gompertz en el mismo grafico
}

