% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/epididact.R
\name{plot_2logis_logito}
\alias{plot_2logis_logito}
\title{plot_2logis_logito}
\usage{
plot_2logis_logito(y0_1, y0_2, r_1, r_2, maxt)
}
\arguments{
\item{y0_1}{Epidemic 1: initial inoculum}

\item{y0_2}{Epidemic 2: initial inoculum}

\item{r_1}{Epidemic 1: rate of progress}

\item{r_2}{Epidemic 2: rate of progress}

\item{maxt}{epidemic duration (days)}
}
\value{
Dos logits rects
}
\description{
Comparison two logits rects in the same plot
}

