% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/epididact.R
\name{plot3logis}
\alias{plot3logis}
\title{plot3logis}
\usage{
plot3logis(y0_1, y0_2, y0_3, r_1, r_2, r_3, maxt)
}
\arguments{
\item{y0_1}{Epidemic 1: initial inoculum}

\item{y0_2}{Epidemic 2: initial inoculum}

\item{y0_3}{Epidemic 3: initial inoculum}

\item{r_1}{Epidemic 1: rate of progress}

\item{r_2}{Epidemic 2: rate of progress}

\item{maxt}{epidemic duration (days)}

\item{r_1}{Epidemic 3: rate of progress}
}
\value{
Three logistic curves
}
\description{
Comparison of three epidemics (logisic model) at the same plot
}
\examples{
plot3logis(0.01, 0.01, 0.01,  # Initial inoculum: Epid 1, Epid 2, Epid 3
           0.05, 0.025, 0.01, # Rate of progress: Epid 1, Epid 2, Epid 3
           100)               # Epidemic duration
}

